pyproject.toml
[tool.poetry]
name = "graduate-work"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pip = ">=24.2,<25.0"
attrs = ">=24.2.0,<24.3.0"
pillow = ">=11.0.0,<11.1.0"
django = ">=5.1.2,<5.2.0"
asgiref = ">=3.8.1,<3.9.0"
sqlparse = ">=0.5.1,<0.6.0"
colorama = ">=0.4.6,<0.5.0"
requests = ">=2.32.3,<2.33.0"
typing-extensions = ">=4.12.2,<4.13.0"
djangorestframework = ">=3.15.2,<3.16.0"
inflection = ">=0.5.1,<0.6.0"
python-dotenv = ">=1.0.1,<1.1.0"
djangorestframework-simplejwt = "^5.3.1"
psycopg2-binary = "^2.9.10"
django-extensions = "^3.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
